pyproject.toml
[tool.poetry]
name = "RT-DETR-Tensorleap"
version = "1"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
code-loader = "1.0.171"
matplotlib = ">=3.3.0"
numpy = ">=1.22.2"
opencv-python = ">=4.6.0"
pillow = ">=7.1.2"
pyyaml = ">=5.3.1"
requests = ">=2.23.0"
scipy = ">=1.4.1"
torch = "2.0.1"
torchvision = "0.15.2"
tqdm = ">=4.64.0"
psutil = "5.9.8"
py-cpuinfo = "*"
thop = ">=0.1.1"
pandas = ">=1.1.4"
seaborn = ">=0.11.0"
ultralytics = "8.4.22"
onnx = "1.17.0"
onnxruntime = "1.19.2"
setuptools = "^82.0.0"
pycocotools = "^2.0.11"
transformers = "4.50.0"
tensorboard = "^2.20.0"
faster-coco-eval = "^1.7.2"
boto3 = "^1.42.91"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
